Update 13_adult.do
generate a_inpatient_ref; 
recode a_inpatient_1y for Philippines2013: according to the questionnaire P342-346, sh208a & sh222a list out the line number of hh members who confined in a hospital. For most of the cases, the individual samples who are assigned with value in sh208a & sh222a are not the members who confined in a hospital. So I generate a_inpatient_1y basing on the line number listed in sh208a & sh222a.